# src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: Xira
  version: 1.0.0
  description: 'Xira - Issue tracking and project management'
...
paths:
...
  /projects:
    get:
      tags: [ Projects ]
      operationId: getProjects
      summary: Get projects of the authenticated user
      description: Returns all projects the authenticated user is a member of.
      security:
        - BearerAuth: [ ]
      responses:
        '200':
          description: Projects retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectSummaryResponse'
        '401':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
    post:
      tags: [ Projects ]
      operationId: createProject
      summary: Create a new project
      description: 'Creates a new project owned by the authenticated user.'
      security:
        - BearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created successfully
          headers:
            Location:
              description: URL of the created project resource
              schema:
                type: string
                format: uri
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
  /projects/{key}:
    get:
      tags: [ Projects ]
      operationId: getProjectByKey
      summary: Get project details
      description: 'Returns detailed information about a project including its boards.'
      security:
        - BearerAuth: [ ]
      parameters:
        - name: key
          in: path
          required: true
          description: Project key
          schema:
            type: string
      responses:
        '200':
          description: Project retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetailsResponse'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
    patch:
      tags: [ Projects ]
      operationId: updateProject
      summary: Update a project
      description: 'Updates project name and/or description. Only accessible by project admins.'
      security:
        - BearerAuth: [ ]
      parameters:
        - name: key
          in: path
          required: true
          description: Project key
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: Project updated successfully
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
...
  /projects/{key}/sprints:
    post:
      tags: [ Sprints ]
      operationId: createSprint
      summary: Add a sprint to a project
      description: 'Creates a new sprint for the specified project. Only project members can manage sprints.'
      security:
        - BearerAuth: [ ]
      parameters:
        - name: key
          in: path
          required: true
          description: Project key
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddSprintRequest'
      responses:
        '201':
          description: Sprint created successfully
          headers:
            Location:
              description: URL of the created sprint resource
              schema:
                type: string
                format: uri
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
  /projects/{key}/sprints/{sprintId}:
    patch:
      tags: [ Sprints ]
      operationId: updateSprint
      summary: Update a sprint
      description: 'Updates sprint name and/or goal. Finished sprints cannot be updated.'
      security:
        - BearerAuth: [ ]
      parameters:
        - name: key
          in: path
          required: true
          description: Project key
          schema:
            type: string
        - name: sprintId
          in: path
          required: true
          description: Sprint identifier
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSprintRequest'
      responses:
        '200':
          description: Sprint updated successfully
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
    delete:
      tags: [ Sprints ]
      operationId: deleteSprint
      summary: Delete a sprint
      description: 'Deletes a sprint that is not active or finished. Issues in the sprint are moved back to the backlog.'
      security:
        - BearerAuth: [ ]
      parameters:
        - name: key
          in: path
          required: true
          description: Project key
          schema:
            type: string
        - name: sprintId
          in: path
          required: true
          description: Sprint identifier
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Sprint deleted successfully
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
  /projects/{key}/sprints/{sprintId}/start:
    post:
      tags: [ Sprints ]
      operationId: startSprint
      summary: Start a sprint
      description: 'Starts a planned sprint and marks it as active. Only one active sprint per project is allowed.'
      security:
        - BearerAuth: [ ]
      parameters:
        - name: key
          in: path
          required: true
          description: Project key
          schema:
            type: string
        - name: sprintId
          in: path
          required: true
          description: Sprint identifier
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Sprint started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SprintResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
  /projects/{key}/sprints/{sprintId}/finish:
    post:
      tags: [ Sprints ]
      operationId: finishSprint
      summary: Finish a sprint
      description: 'Finishes an active sprint. Unfinished issues are returned to the backlog while completed issues remain assigned.'
      security:
        - BearerAuth: [ ]
      parameters:
        - name: key
          in: path
          required: true
          description: Project key
          schema:
            type: string
        - name: sprintId
          in: path
          required: true
          description: Sprint identifier
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Sprint finished successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SprintResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
...
  /projects/{key}/members:
    post:
      tags: [ Projects ]
      operationId: addProjectMember
      summary: Add a member to a project
      description: 'Adds a new member to the project. Only accessible by project admins.'
      security:
        - BearerAuth: [ ]
      parameters:
        - name: key
          in: path
          required: true
          description: Project key
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddProjectMemberRequest'
      responses:
        '201':
          description: Project member added successfully
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
...
  /projects/{key}/workflows/statuses:
    get:
      tags: [ Projects ]
      operationId: getWorkflowStatuses
      summary: Get all workflow statuses of a project
      description: 'Returns all workflow statuses for the specified project, sorted by status order.'
      security:
        - BearerAuth: [ ]
      parameters:
        - name: key
          in: path
          required: true
          description: Project key
          schema:
            type: string
      responses:
        '200':
          description: Workflow statuses retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkflowStatusResponse'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
  /projects/{key}/members/{userId}:
    patch:
      tags: [ Projects ]
      operationId: updateProjectMemberRole
      summary: Update the role of a project member
      description: 'Updates the role of an existing project member. Only accessible by project admins.'
      security:
        - BearerAuth: [ ]
      parameters:
        - name: key
          in: path
          required: true
          description: Project key
          schema:
            type: string
        - name: userId
          in: path
          required: true
          description: User ID of the project member
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectMemberRoleRequest'
      responses:
        '200':
          description: Project member role updated successfully
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
    delete:
      tags: [ Projects ]
      operationId: removeProjectMember
      summary: Remove a member from a project
      description: 'Removes an existing member from the project. Only accessible by project admins.'
      security:
        - BearerAuth: [ ]
      parameters:
        - name: key
          in: path
          required: true
          description: Project key
          schema:
            type: string
        - name: userId
          in: path
          required: true
          description: User ID of the project member
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Project member removed successfully
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '500':
          $ref: '#/components/responses/Problem'
components:
  responses:
    Problem:
      description: Default response in case of an error.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ProjectBoardResponse:
      type: object
      additionalProperties: false
      required: [ name, boardNumber, type ]
      properties:
        name:
          type: string
          description: Board name
        boardNumber:
          type: integer
          format: int32
          description: Unique board number within the project
        type:
          type: string
          description: Board type
          enum: [ SCRUM, KANBAN ]
    ProjectSummaryResponse:
      type: object
      additionalProperties: false
      required: [ key, name, isOwner, isAdmin ]
      properties:
        key:
          type: string
          description: Project key
        name:
          type: string
          description: Project name
        description:
          type: string
          description: Project description
          nullable: true
        isOwner:
          type: boolean
          description: Whether the authenticated user is the owner of the project
        isAdmin:
          type: boolean
          description: Whether the authenticated user is an admin of the project
    ProjectDetailsResponse:
      allOf:
        - $ref: '#/components/schemas/ProjectSummaryResponse'
        - type: object
          required: [ boards ]
          properties:
            boards:
              type: array
              items:
                $ref: '#/components/schemas/ProjectBoardResponse'
    Problem:
      type: object
      additionalProperties: true
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
...
    AddProjectMemberRequest:
      type: object
      additionalProperties: false
      required: [ userId, role ]
      properties:
        userId:
          type: string
          format: uuid
          description: User ID of the user to add as project member
        role:
          $ref: '#/components/schemas/ProjectMemberRole'
    ProjectMemberRole:
      type: string
      description: Role of a user within a project
      enum: [ ADMIN, DEVELOPER ]
    CreateProjectRequest:
      type: object
      additionalProperties: false
      required: [ key, name ]
      properties:
        key:
          type: string
          description: Unique project key (e.g. XIRA)
          minLength: 2
          maxLength: 10
          pattern: '[A-Z]*$'
        name:
          type: string
          description: Human-readable project name
          minLength: 1
          maxLength: 100
        description:
          type: string
          description: Optional project description
          maxLength: 200
    UpdateProjectRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          description: Human-readable project name
          minLength: 1
          maxLength: 100
        description:
          type: string
          description: Project description
          minLength: 1
          maxLength: 200
...
    AddSprintRequest:
      type: object
      additionalProperties: false
      required: [ name, goal ]
      properties:
        name:
          type: string
          description: Sprint name
          minLength: 1
          maxLength: 100
        goal:
          type: string
          description: Sprint goal
          minLength: 1
          maxLength: 200
    UpdateSprintRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          description: Sprint name
          minLength: 1
          maxLength: 100
        goal:
          type: string
          description: Sprint goal
          minLength: 1
          maxLength: 200
    SprintResponse:
      type: object
      additionalProperties: false
      required: [ id, name, goal, state, createdAt ]
      properties:
        id:
          type: string
          format: uuid
          description: Sprint identifier
        name:
          type: string
          description: Sprint name
        goal:
          type: string
          description: Sprint goal
        state:
          $ref: '#/components/schemas/SprintState'
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the sprint was created
        startedAt:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when the sprint was started
        finishedAt:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when the sprint was finished
    SprintState:
      type: string
      description: Lifecycle state of a sprint
      enum: [ PLANNED, ACTIVE, CLOSED ]
    UpdateProjectMemberRoleRequest:
      type: object
      additionalProperties: false
      required: [ role ]
      properties:
        role:
          $ref: '#/components/schemas/ProjectMemberRole'
...
    WorkflowStatusResponse:
      type: object
      additionalProperties: false
      required: [ id, name, category ]
      properties:
        id:
          type: string
          format: uuid
          description: Workflow status identifier
        name:
          type: string
          description: Workflow status name
        category:
          type: string
          description: Workflow status category
          enum: [ TODO, IN_PROGRESS, DONE ]
...